Add Bassman Ultra
This audio configuration offers a significantly amplified bass output, surpassing the levels provided by the Bassman preset file. It is meticulously designed for aficionados of profound and resonant bass who are unwilling to compromise on the minutest of details. 

The experience it delivers is nothing short of extraordinary, producing a bass output so powerful that it could metaphorically shake one's brain. It is an absolute blessing for your In-Ear Monitors (IEMs), enhancing their performance to deliver an auditory experience that is unparalleled and beyond anything you have ever encountered. 

In essence, this configuration is a testament to our commitment to cater to the true bass enthusiasts, ensuring they enjoy every beat to its fullest. Experience the magic of sound like never before. Enjoy the rhythm, feel the bass.